#30 In Progress Changes: - Add command `mtasa-lua-utils` with subcommands: `boilerplate` and `build`
Caveats:
- `ts.sys.args` contains 1 argument
- `parseCommandLine(...)` in `buildProject` should accept sliced args array
